move new post out of post successfully & clean up
