+++ src/components/input.js
@@ -18,7 +18,7 @@ const SetupCommunityInput = styled(
 
 const PostInput = styled(
     TextInput,
-    'w-full border-none rounded-lg text-sm focus:text-base bg-orchid-100 p-3 placeholder-orchid-300 text-orchid-900'
+    'w-full border-none rounded-lg text-sm bg-orchid-100 p-3 placeholder-orchid-300 text-orchid-900'
 )
 
 export { LoginInput, RegisterInput, SetupCommunityInput, PostInput }
