+++ src/screens/community/Post.js
@@ -5,32 +5,14 @@ import axios from 'axios'
 import { BASE_URL } from '@env'
 
 import AsyncStorage from '@react-native-async-storage/async-storage'
-import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome'
-import {
-    faCircleXmark,
-    faPlus,
-    faPencil,
-    faMagnifyingGlass,
-} from '@fortawesome/free-solid-svg-icons'
 
 import Ionicons from 'react-native-vector-icons/Ionicons'
-import { PostInput } from '../../components/input'
-import STRINGS from '../../constants/strings'
-import SIZES from '../../constants/sizes'
-
-import { uploadImage } from '../../utils/fbHelper'
 
 import Modal from 'react-native-modal'
 import COLORS from '../../constants/colors'
 import PostItem from './components/PostItem'
-
-import * as Progress from 'react-native-progress'
-
-import * as ImagePicker from 'expo-image-picker'
 import { useRoute } from '@react-navigation/native'
 import ImageView from 'react-native-image-viewing'
-
-import { getDatetime } from '../../utils/Utils'
 
 import TopicSelector from './components/TopicSelector'
 
@@ -38,14 +20,11 @@ import {
     View,
     Text,
     TouchableOpacity,
-    Image,
     ScrollView,
     RefreshControl,
-    Alert,
 } from 'react-native'
-import { PostButton } from '../../components/button'
-import { SearchBar } from 'react-native-elements'
-import { inputStyle, searchBarStyle } from '../../../styles/style'
+
+import NewPost from './components/NewPost'
 
 const Post = (nav) => {
     const [postContent, setPostContent] = useState('')
@@ -59,25 +38,17 @@ const Post = (nav) => {
     )
     const [pickedImages, setPickedImages] = useState([])
 
-    const [uploadProgress, setUploadProgress] = useState(0)
-
-    const [eachImageProgress, setEachImageProgress] = useState([])
-
-    const [imageURLs, setImageURLs] = useState([])
-
     const [refreshing, setRefreshing] = React.useState(false)
 
     const [isJoined, setIsJoined] = useState(useRoute().params.communityId)
 
+    const [openTopicSelector, setOpenTopicSelector] = useState(false)
+
     const [imageViewerVisible, setImageViewerVisible] = useState(false)
 
-    const [tagUserSearchBarVisibility, setTagUserSearchBarVisibility] =
-        useState(false)
-
-    const [isTopicSelectorModalVisible, setIsTopicSelectorModalVisible] =
-        useState(false)
-
     const [selectedTopic, setSelectedTopic] = useState({})
+
+    const [isNewData, setIsNewData] = useState(false)
 
     const onRefresh = React.useCallback(() => {
         setRefreshing(true)
@@ -86,44 +57,8 @@ const Post = (nav) => {
             setRefreshing(false)
         }, 1000)
     }, [])
-    const [searchUsername, setSearchUsername] = useState('')
-    const [searchUserList, setSearchUserList] = useState([])
+
     const [taggedUsers, setTaggedUsers] = useState([])
-
-    const [isPosting, setIsPosting] = useState(false)
-
-    const tagUser = (index) => {
-        setTaggedUsers([...taggedUsers, searchUserList[index]])
-        searchUserList.splice(index, 1)
-        setSearchUsername('')
-    }
-
-    const removeTaggedUser = (index) => {
-        const removedUser = taggedUsers.slice(index)
-        const removeUserId = taggedUsersId.slice(index)
-
-        setTaggedUsers(
-            taggedUsers.filter((user) => user.id !== removedUser[0].id)
-        )
-        setTaggedUsersId(taggedUsersId.filter((id) => id !== removeUserId[0]))
-
-        if (removedUser[0].username.includes(searchUsername.toLowerCase())) {
-            setSearchUserList([...searchUserList, removedUser[0]])
-        }
-    }
-
-    const [taggedUsersId, setTaggedUsersId] = useState([])
-
-    useEffect(() => {
-        // console.log('Tagged users:', taggedUsersId)
-        // console.log('Tagged users:', taggedUsers)
-    }, [taggedUsersId])
-
-    useEffect(() => {
-        taggedUsers.map((user) => {
-            setTaggedUsersId([...taggedUsersId, user.id])
-        })
-    }, [taggedUsers])
 
     const findUserId = async () => {
         try {
@@ -136,39 +71,18 @@ const Post = (nav) => {
         }
     }
 
-    const pickingImageHandler = async () => {
-        let result = await ImagePicker.launchImageLibraryAsync({
-            mediaTypes: ImagePicker.MediaTypeOptions.Images,
-            allowsEditing: false,
-            quality: 0.8,
-        })
-
-        // console.log(result)
-
-        if (!result.canceled) {
-            const resultImage = result.assets[0]
-            setPickedImages([...pickedImages, resultImage])
-            setEachImageProgress(
-                eachImageProgress.concat(Array(pickedImages.length).fill(0))
-            )
-        } else {
-            // alert('You did not select any image.')
-        }
-    }
-
-    useEffect(() => {
-        // console.log('Picked images:', pickedImages.length)
-        setUploadProgress(0)
-        setEachImageProgress([])
-    }, [pickedImages])
-
     useEffect(() => {
         getPosts()
         findUserId()
-
-        // console.log(community_id)
-        // console.log(userId)
     }, [])
+
+    useEffect(() => {
+        if (isNewData) {
+            console.log('new data')
+            getPosts()
+            setIsNewData(false)
+        }
+    }, [isNewData])
 
     const getPosts = async () => {
         axios({
@@ -179,8 +93,6 @@ const Post = (nav) => {
             },
         })
             .then((res) => {
-                // reverse the array to show the latest post first
-                // console.log('Posts:', res.data)
                 setPostData(res.data.reverse())
             })
             .catch((err) => {
@@ -190,170 +102,6 @@ const Post = (nav) => {
 
     const handleOpenPopup = () => {
         setIsModalVisible(true)
-    }
-
-    const handleCancelPopup = () => {
-        Alert.alert(
-            'Discard Post',
-            'Are you sure you want to discard this post? You will lose all inputs.',
-            [
-                {
-                    text: 'No',
-                    onPress: () => console.log('Cancel Pressed'),
-                    style: 'cancel',
-                },
-                {
-                    text: 'Yes',
-                    onPress: () => {
-                        setPostContent('')
-                        setPickedImages([])
-                        setImageURLs([])
-                        setTaggedUsers([])
-                        setTaggedUsersId([])
-                        setIsModalVisible(false)
-                        setIsTopicSelectorModalVisible(false)
-                        setSelectedTopic({})
-                    },
-                },
-            ],
-            { cancelable: false }
-        )
-    }
-
-    const handleClosePopup = () => {
-        setPostContent('')
-        setPickedImages([])
-        setImageURLs([])
-        setTaggedUsers([])
-        setTaggedUsersId([])
-        setIsModalVisible(false)
-        setIsTopicSelectorModalVisible(false)
-        setSelectedTopic({})
-    }
-
-    useEffect(() => {
-        // console.log('Each image progress:', eachImageProgress)
-        if (eachImageProgress.length > 0) {
-            setUploadProgress(
-                eachImageProgress.reduce((a, b) => a + b, 0) /
-                    eachImageProgress.length /
-                    100
-            )
-        }
-    }, [eachImageProgress])
-
-    // SUCCESS MESSAGE HERE
-    useEffect(() => {
-        if (
-            imageURLs.length != 0 &&
-            pickedImages != 0 &&
-            imageURLs.length === pickedImages.length
-        ) {
-            // console.log('Image URLs:', imageURLs)
-
-            setPickedImages([])
-            setEachImageProgress([])
-
-            // turn array of objects into array of strings
-
-            axios({
-                method: 'POST',
-                url: `${BASE_URL}/post/create`,
-                data: {
-                    text: postContent,
-                    communityId: community_id,
-                    authorId: userId,
-                    dateTime: getDatetime(),
-                    medias: JSON.stringify({ urls: imageURLs }),
-                    //join the tagged users id into a string
-                    tagged: taggedUsersId.join(','),
-                    topicId: selectedTopic.id,
-                },
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-            })
-                .then((res) => {
-                    // console.log('Post published', res.data)
-
-                    setPostData(res.data)
-                    handleClosePopup()
-                })
-                .catch((err) => {
-                    setIsPosting(false)
-                    console.log('Cannot publish the post', err)
-                })
-        }
-    }, [imageURLs])
-
-    const handlePublish = async () => {
-        setIsPosting(true)
-
-        setImageURLs([])
-
-        for (let i = 0; i < pickedImages.length; i++) {
-            try {
-                const uploadedImage = await uploadImage(
-                    community_id,
-                    userId,
-                    pickedImages[i],
-                    i,
-                    setEachImageProgress,
-                    setImageURLs
-                )
-            } catch (e) {
-                setIsPosting(false)
-                console.log(e)
-                return
-            }
-        }
-    }
-
-    const removePickedImage = (index) => {
-        pickedImages.splice(index, 1)
-        setPickedImages([...pickedImages])
-    }
-
-    useEffect(() => {
-        axios({
-            method: 'GET',
-            url: `${BASE_URL}/user/searchUsersInCommunity?filter=${searchUsername}&communityId=${community_id}`,
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        })
-            .then((res) => {
-                // console.log('Search user:', res.data)
-                const filteredUsers = res.data.filter(
-                    (user) =>
-                        !taggedUsers.some(
-                            (taggedUser) => taggedUser.id === user.id
-                        )
-                )
-                setSearchUserList(filteredUsers)
-            })
-            .catch((err) => {
-                console.log(err)
-            })
-    }, [searchUsername])
-
-    const searchUser = (text) => {
-        setSearchUsername(text)
-    }
-
-    const handleAddTag = () => {
-        //
-    }
-
-    const handleEditTopic = () => {
-        setIsModalVisible(false)
-        setIsTopicSelectorModalVisible(true)
-    }
-
-    const [search, setSearch] = useState('')
-
-    const updateSearch = (search) => {
-        setSearch(search)
     }
 
     return (
@@ -412,377 +160,39 @@ const Post = (nav) => {
                         animationType="slide"
                         presentationStyle="pageSheet"
                     >
-                        <ScrollView
-                            className="flex-1"
-                            showsVerticalScrollIndicator={false}
-                            pointerEvents={isPosting ? 'none' : 'auto'}
-                        >
-                            <View className="flex flex-1">
-                                <View className="w-fit flex-col items-center justify-start space-y-5">
-                                    <View className="flex w-full items-center">
-                                        <Text className="text-2xl font-bold text-orchid-900">
-                                            {STRINGS.CreatePost}
-                                        </Text>
-                                    </View>
-                                    <PostInput
-                                        className="h-48"
-                                        multiline={true}
-                                        placeholder={
-                                            STRINGS.postContentPlaceholder
-                                        }
-                                        value={postContent}
-                                        onChangeText={(text) =>
-                                            setPostContent(text)
-                                        }
-                                    />
-
-                                    <View className="flex h-fit w-full flex-row items-center justify-between">
-                                        <Text className="text-orchid-900">
-                                            Select a Post Topic
-                                        </Text>
-                                        <View className="flex-row items-center justify-center">
-                                            <TouchableOpacity
-                                                onPress={() =>
-                                                    handleEditTopic()
-                                                }
-                                                className="flex-row items-center justify-center space-x-2 rounded-full bg-gold-900 px-6 py-2"
-                                            >
-                                                {selectedTopic.name && (
-                                                    <Text className="text-orchid-900">
-                                                        {selectedTopic.name}
-                                                    </Text>
-                                                )}
-                                                <FontAwesomeIcon
-                                                    icon={faPencil}
-                                                    color={COLORS.orchid[900]}
-                                                    size={SIZES.smallIconSize}
-                                                />
-                                            </TouchableOpacity>
-                                        </View>
-                                    </View>
-
-                                    <View className="flex h-fit w-full">
-                                        <View className="mb-2 flex flex-row items-center justify-between">
-                                            <Text className="text-orchid-900">
-                                                Tag Users
-                                            </Text>
-                                            <TouchableOpacity
-                                                onPress={() =>
-                                                    setTagUserSearchBarVisibility(
-                                                        (prev) => !prev
-                                                    )
-                                                }
-                                                className="flex-row items-center justify-center space-x-2 rounded-full bg-gold-900 px-5 py-2"
-                                            >
-                                                <Text className="text-sm text-orchid-900">
-                                                    Find Users
-                                                </Text>
-                                                <FontAwesomeIcon
-                                                    icon={faMagnifyingGlass}
-                                                    color={COLORS.orchid[900]}
-                                                    size={SIZES.smallIconSize}
-                                                />
-                                            </TouchableOpacity>
-                                        </View>
-                                        {tagUserSearchBarVisibility && (
-                                            <View className="w-fit flex-col items-center justify-start">
-                                                <SearchBar
-                                                    placeholder={
-                                                        STRINGS.TagUsers
-                                                    }
-                                                    onChangeText={(text) =>
-                                                        searchUser(text)
-                                                    }
-                                                    value={searchUsername}
-                                                    containerStyle={[
-                                                        searchBarStyle.containerSearchBar,
-                                                        inputStyle.inputShadow,
-                                                    ]}
-                                                    inputContainerStyle={
-                                                        searchBarStyle.inputSearchBar
-                                                    }
-                                                    inputStyle={
-                                                        searchBarStyle.input
-                                                    }
-                                                    placeholderTextColor={
-                                                        COLORS['orchid'][400]
-                                                    }
-                                                    searchIcon={
-                                                        searchBarStyle.seachIcon
-                                                    }
-                                                    clearIcon={
-                                                        searchBarStyle.clearIcon
-                                                    }
-                                                />
-
-                                                {searchUsername.length > 0 &&
-                                                    searchUserList.length >
-                                                        0 && (
-                                                        <ScrollView
-                                                            className="mt-1 h-24 w-fit overflow-auto"
-                                                            contentContainerStyle={{
-                                                                flexDirection:
-                                                                    'row',
-                                                                flexGrow: 1,
-                                                            }}
-                                                        >
-                                                            <View className="flex h-full w-full flex-row flex-wrap items-start justify-start overflow-auto">
-                                                                {searchUserList.map(
-                                                                    (
-                                                                        user,
-                                                                        index
-                                                                    ) => (
-                                                                        <TouchableOpacity
-                                                                            key={
-                                                                                index
-                                                                            }
-                                                                            className="mx-1 my-2 rounded-full bg-orchid-100 px-3 py-1.5 shadow-sm"
-                                                                            onPress={() =>
-                                                                                tagUser(
-                                                                                    index
-                                                                                )
-                                                                            }
-                                                                        >
-                                                                            <Text className="text-sm text-orchid-900">
-                                                                                {
-                                                                                    user.username
-                                                                                }
-                                                                            </Text>
-                                                                        </TouchableOpacity>
-                                                                    )
-                                                                )}
-                                                            </View>
-                                                        </ScrollView>
-                                                    )}
-                                            </View>
-                                        )}
-                                    </View>
-
-                                    {taggedUsers.length > 0 && (
-                                        <View className="w-full space-y-1">
-                                            <View className="flex w-full flex-col items-start justify-start">
-                                                <Text className="text-orchid-900">
-                                                    Tagged Users
-                                                </Text>
-                                            </View>
-
-                                            <ScrollView
-                                                className="min-h-16 max-h-26 w-fit overflow-auto"
-                                                contentContainerStyle={{
-                                                    display: 'flex',
-                                                    flexDirection: 'row',
-                                                    flexWrap: 'wrap',
-                                                    justifyContent:
-                                                        'flex-start',
-                                                    flexGrow: 1,
-                                                }}
-                                            >
-                                                {taggedUsers.map(
-                                                    (user, index) => (
-                                                        <View className="mx-1 my-2 flex flex-row items-center justify-center space-x-2 rounded-full bg-orchid-100 px-3 py-1.5 shadow-sm">
-                                                            <View>
-                                                                <Text className="text-sm text-orchid-900">
-                                                                    {
-                                                                        user.username
-                                                                    }
-                                                                </Text>
-                                                            </View>
-                                                            <TouchableOpacity
-                                                                key={index}
-                                                                onPress={() =>
-                                                                    removeTaggedUser(
-                                                                        index
-                                                                    )
-                                                                }
-                                                            >
-                                                                <FontAwesomeIcon
-                                                                    icon={
-                                                                        faCircleXmark
-                                                                    }
-                                                                    color={
-                                                                        COLORS[
-                                                                            'orchid'
-                                                                        ][800]
-                                                                    }
-                                                                    size={
-                                                                        SIZES[
-                                                                            'xMarkIconSizeTag'
-                                                                        ]
-                                                                    }
-                                                                />
-                                                            </TouchableOpacity>
-                                                        </View>
-                                                    )
-                                                )}
-                                            </ScrollView>
-                                        </View>
-                                    )}
-
-                                    <View className="h-fit w-full flex-row items-center justify-center space-x-2">
-                                        {pickedImages.length >= 0 &&
-                                            pickedImages.length < 5 && (
-                                                <TouchableOpacity
-                                                    onPress={() =>
-                                                        pickingImageHandler()
-                                                    }
-                                                    className="flex h-16 w-16 items-center justify-center space-y-1 rounded-lg bg-orchid-100"
-                                                >
-                                                    <FontAwesomeIcon
-                                                        icon={faPlus}
-                                                        color={
-                                                            COLORS.orchid[900]
-                                                        }
-                                                        size={
-                                                            SIZES.postImageIconSize
-                                                        }
-                                                    />
-                                                    <Text className="text-xs text-orchid-900">
-                                                        Images
-                                                    </Text>
-                                                </TouchableOpacity>
-                                            )}
-                                        {pickedImages.length > 0 &&
-                                            pickedImages.map(
-                                                (imageInfo, index) => {
-                                                    return (
-                                                        <TouchableOpacity
-                                                            onPress={() =>
-                                                                setImageViewerVisible(
-                                                                    true
-                                                                )
-                                                            }
-                                                        >
-                                                            <View className="relative h-16 w-16 bg-transparent">
-                                                                <Image
-                                                                    source={{
-                                                                        uri: imageInfo.uri,
-                                                                    }}
-                                                                    className="h-16 w-16 rounded-lg"
-                                                                />
-                                                                <TouchableOpacity
-                                                                    className="absolute right-1 top-1"
-                                                                    onPress={() =>
-                                                                        removePickedImage(
-                                                                            index
-                                                                        )
-                                                                    }
-                                                                >
-                                                                    <FontAwesomeIcon
-                                                                        icon={
-                                                                            faCircleXmark
-                                                                        }
-                                                                        color={
-                                                                            COLORS.white
-                                                                        }
-                                                                    />
-                                                                </TouchableOpacity>
-                                                            </View>
-                                                        </TouchableOpacity>
-                                                    )
-                                                }
-                                            )}
-                                        {pickedImages.length < 5 &&
-                                            Array.from(
-                                                {
-                                                    length:
-                                                        4 - pickedImages.length,
-                                                },
-                                                (_, i) => (
-                                                    <View className="h-16 w-16 rounded-lg bg-slate-100"></View>
-                                                )
-                                            )}
-                                    </View>
-                                    {uploadProgress > 0 &&
-                                        uploadProgress < 1 && (
-                                            <View className="flex h-fit w-4/5 flex-col items-center justify-center space-y-1">
-                                                <View className="flex h-2 w-full shadow shadow-slate-200">
-                                                    <Progress.Bar
-                                                        animated={true}
-                                                        progress={
-                                                            uploadProgress
-                                                        }
-                                                        width={null}
-                                                        height={6}
-                                                        borderRadius={4}
-                                                        borderColor={
-                                                            COLORS.orchid[500]
-                                                        }
-                                                        borderWidth={0}
-                                                        unfilledColor={
-                                                            COLORS.orchid[100]
-                                                        }
-                                                        color={
-                                                            COLORS.orchid[500]
-                                                        }
-                                                    />
-                                                </View>
-                                                <Text>
-                                                    {Math.round(
-                                                        uploadProgress * 100
-                                                    )}
-                                                    %
-                                                </Text>
-                                            </View>
-                                        )}
-                                    {uploadProgress === 1 && (
-                                        <View className="flex h-fit w-4/5 flex-col items-center justify-center space-y-1">
-                                            <Text className="text-orchid-900">
-                                                {STRINGS.uploadSuccess}
-                                            </Text>
-                                        </View>
-                                    )}
-                                    {/* {uploadProgress === 0 && (
-                                        <View className="flex h-fit w-4/5 flex-col items-center justify-center space-y-1">
-                                            <Text className="text-orchid-900">
-                                                {pickedImages.length}/5 images
-                                                selected
-                                            </Text>
-                                        </View>
-                                    )} */}
-                                    <View className="flex-row justify-evenly space-x-10">
-                                        <PostButton
-                                            onPress={() => handlePublish()}
-                                            className="bg-gold-900"
-                                        >
-                                            <Text className="text-base text-orchid-900">
-                                                {' '}
-                                                {STRINGS.publish}
-                                            </Text>
-                                        </PostButton>
-                                        <PostButton
-                                            onPress={() => handleCancelPopup()}
-                                            className="bg-gray-300"
-                                        >
-                                            <Text className="justify-center text-base  text-orchid-900">
-                                                {' '}
-                                                {STRINGS.cancel}
-                                            </Text>
-                                        </PostButton>
-                                    </View>
-                                </View>
-                            </View>
-                        </ScrollView>
-                    </Modal>
-
-                    <Modal
-                        presentationStyle="pageSheet"
-                        visible={
-                            isTopicSelectorModalVisible &&
-                            !imageViewerVisible &&
-                            !IsModalVisible
-                        }
-                        transparent={false}
-                        animationType="slide"
-                    >
-                        {/* safe area? */}
-
-                        <TopicSelector
-                            key={Math.random()}
-                            setModalVisibility={setIsTopicSelectorModalVisible}
-                            setPostModalVisibility={setIsModalVisible}
-                            selectedTopic={selectedTopic}
-                            setSelectedTopic={setSelectedTopic}
-                        />
+                        {openTopicSelector ? (
+                            <TopicSelector
+                                key={Math.random()}
+                                openTopicSelector={openTopicSelector}
+                                setOpenTopicSelector={setOpenTopicSelector}
+                                selectedTopic={selectedTopic}
+                                setSelectedTopic={setSelectedTopic}
+                            />
+                        ) : (
+                            <ScrollView
+                                className="flex-1"
+                                showsVerticalScrollIndicator={false}
+                            >
+                                <NewPost
+                                    userId={userId}
+                                    community_id={community_id}
+                                    isPostModalVisible={IsModalVisible}
+                                    setIsPostModalVisible={setIsModalVisible}
+                                    openTopicSelector={openTopicSelector}
+                                    setOpenTopicSelector={setOpenTopicSelector}
+                                    selectedTopic={selectedTopic}
+                                    setSelectedTopic={setSelectedTopic}
+                                    postContent={postContent}
+                                    setPostContent={setPostContent}
+                                    taggedUsers={taggedUsers}
+                                    setTaggedUsers={setTaggedUsers}
+                                    pImages={pickedImages}
+                                    setPImages={setPickedImages}
+                                    isNewData={isNewData}
+                                    setIsNewData={setIsNewData}
+                                />
+                            </ScrollView>
+                        )}
                     </Modal>
                 </View>
             </View>
