+++ src/screens/community/components/TopicSelector.js
@@ -9,13 +9,13 @@ import { ScrollView } from 'react-native-gesture-handler'
 import STRINGS from '../../../constants/strings'
 
 export default TopicSelector = ({
-    setModalVisibility,
-    setPostModalVisibility,
+    openTopicSelector,
+    setOpenTopicSelector,
     selectedTopic,
     setSelectedTopic,
 }) => {
     const [topicList, setTopicList] = useState({})
-    const [currentTopic, setCurrentTopic] = useState({selectedTopic})
+    const [currentTopic, setCurrentTopic] = useState({ selectedTopic })
 
     const getTopics = async (name) => {
         axios({
@@ -38,7 +38,6 @@ export default TopicSelector = ({
     }, [])
 
     useEffect(() => {
-        // console.log('Topic List:', topicList)
     }, [topicList])
 
     return (
@@ -73,7 +72,6 @@ export default TopicSelector = ({
                                     name: topicList[key],
                                 })
                             }}
-
                         >
                             <Text className="text-sm text-orchid-900">
                                 {topicList[key]}
@@ -86,8 +84,7 @@ export default TopicSelector = ({
                 <TouchableOpacity
                     className="flex h-fit flex-1 items-center rounded-xl bg-red-300 px-8 py-2 shadow"
                     onPress={() => {
-                        setPostModalVisibility(true)
-                        setModalVisibility(false)
+                        setOpenTopicSelector((prev) => !prev)
                     }}
                 >
                     <Text className="align-middle text-base font-bold text-red-900">
@@ -98,8 +95,7 @@ export default TopicSelector = ({
                     className="flex h-fit w-1/2 items-center rounded-xl bg-green-300 px-10 py-2 shadow"
                     onPress={() => {
                         setSelectedTopic(currentTopic)
-                        setModalVisibility(false)
-                        setPostModalVisibility(true)
+                        setOpenTopicSelector((prev) => !prev)
                     }}
                 >
                     <Text className="align-middle text-base font-bold text-green-900">
